refactor(cockpit): break cockpit_inbox <-> cockpit_inbox_items cycle (refs #1367)
Breaks the `pollypm.cockpit_inbox <-> pollypm.cockpit_inbox_items` circular
import pair from #1367. Extracts the two leaf helpers `_inbox_db_sources`
and `_row_is_dev_channel` (no pollypm deps; pure stdlib) into a new
`pollypm.cockpit_inbox_sources` module. Both sides now import from the
leaf; the top-level `cockpit_inbox_items -> cockpit_inbox` edge is gone.

- `cockpit_inbox` re-exports both helpers via `__all__` so existing
  callers (`cockpit.py` re-export, tests importing from
  `pollypm.cockpit_inbox`) keep working unchanged.
- `cockpit_ui._get_dashboard_inbox_data` redirected to the leaf so a
  lazy-import workaround no longer crosses the (former) cycle.

The lazy `cockpit_inbox_items` imports inside `pm_inbox_awaits_user_list`
and `pm_inbox_kind_list` stay — the cycle they served is gone, but
hoisting them eagerly triggers a separate `work.sqlite_service` import-
ordering issue tracked under #1367. That's a different cluster.

Partial slice — issue stays open. Pair count in scan drops from 29 to 28.
