+++ src/pollypm/cockpit_inbox.py
@@ -29,6 +29,10 @@ from datetime import UTC, datetime, timedelta
 from pathlib import Path
 from typing import Literal
 
+from pollypm.cockpit_inbox_sources import (
+    _inbox_db_sources,
+    _row_is_dev_channel,
+)
 from pollypm.config import load_config
 from pollypm.cockpit_task_priority import priority_glyph, priority_rank
 from pollypm.cockpit_sections.base import _STATUS_ICONS
@@ -38,6 +42,12 @@ from pollypm.cockpit_worker_identity import (
 )
 from pollypm.heartbeats.snapshots import read_recent_heartbeat_snapshot
 from pollypm.inbox_message_refs import row_project_refs as _row_project_refs
+
+__all__ = (
+    # Re-exported leaf helpers (now live in pollypm.cockpit_inbox_sources)
+    "_inbox_db_sources",
+    "_row_is_dev_channel",
+)
 
 
 def _timestamp_sort_value(value) -> float:
@@ -112,68 +122,6 @@ def _render_work_service_issues(project: object) -> str:
         lines.append("No tasks found.")
 
     return "\n".join(lines)
-
-
-def _inbox_db_sources(config) -> list[tuple[str | None, Path, Path]]:
-    """Return ``(project_key, db_path, project_path)`` for every inbox source.
-
-    Includes the per-project ``.pollypm/state.db`` for every registered
-    project **and** the workspace-root ``<workspace_root>/.pollypm/state.db``
-    — the latter is where ``pm notify`` (with defaults) lands items that
-    don't belong to any one project (#271). Duplicates are dropped so a
-    project whose path happens to equal the workspace root is scanned once.
-
-    Note: this helper INCLUDES non-tracked projects on purpose — the
-    inbox cockpit panel and rail badge surface all registered projects
-    so a user can triage messages even before running ``pm init-tracker``.
-    The recovery prompt + morning-briefing surfaces filter to tracked
-    only; they have a different invariant (cycles 85/86/87).
-
-    ``project_key`` is ``None`` for the workspace-root source; callers that
-    need a project filter treat ``None`` as "no filter".
-    """
-    sources: list[tuple[str | None, Path, Path]] = []
-    seen: set[Path] = set()
-    for project_key, project in getattr(config, "projects", {}).items():
-        project_path = Path(project.path)
-        db_path = project_path / ".pollypm" / "state.db"
-        resolved = db_path.resolve() if db_path.exists() else db_path
-        if resolved in seen:
-            continue
-        seen.add(resolved)
-        sources.append((project_key, db_path, project_path))
-
-    workspace_root = getattr(getattr(config, "project", None), "workspace_root", None)
-    if workspace_root is not None:
-        ws_path = Path(workspace_root)
-        ws_db = ws_path / ".pollypm" / "state.db"
-        resolved = ws_db.resolve() if ws_db.exists() else ws_db
-        if resolved not in seen:
-            seen.add(resolved)
-            sources.append((None, ws_db, ws_path))
-    return sources
-
-
-def _row_is_dev_channel(labels_raw: object) -> bool:
-    """Return True if the message's labels list contains ``channel:dev``.
-
-    Accepts the raw column value (JSON string or list) since
-    ``Store.query_messages`` may surface either depending on the
-    engine. Any other channel (or no explicit channel) is treated as
-    user-facing. See #754.
-    """
-    import json as _json
-    labels: list[str] = []
-    if isinstance(labels_raw, list):
-        labels = [str(x) for x in labels_raw]
-    elif isinstance(labels_raw, str) and labels_raw:
-        try:
-            parsed = _json.loads(labels_raw)
-            if isinstance(parsed, list):
-                labels = [str(x) for x in parsed]
-        except ValueError:
-            labels = []
-    return "channel:dev" in labels
 
 
 def pm_inbox_awaits_user_list(config) -> list[object]:
